requirements.txt
mysql.connector
PySimpleGUI